refactor: move assertions from Check.TypeAssertions.cs into their own files
